Modificacion del tipo de dato de Lng y Lat

+++ database/migrations/2016_03_29_180541_pantallas_migration.php
@@ -18,9 +18,9 @@ class PantallasMigration extends Migration
             // Indicamos cual es la clave foránea de esta tabla:
             $table->foreign('Categoria_id')->references('IdCategoria')->on('categorias');
             $table->string('Ubicacion');
-            $table->integer('Lat');
-            $table->integer('Lng');
-            
+            $table->decimal('Lat',5,6);
+            $table->decimal('Lng',5,6);
+
             $table->timestamps();
         });
     }
